implemented font code and keyboard string display
Using java graphics 2d library I implemented anti aliasing into the user interface. This makes it so that the fonts are not displayed pixelated.

I am using the same font family at the moment and resizing it using the deriveFont method, and it is working for now.

Surprisingly, the GameMaker code is looking pretty similar to this code.

I managed to use Key Listener to implement a clone of GameMaker's keyboard_string built in variable. In Game Maker keyboard_string, stores the last 1024 chars the user has typed in a string. This was helpful for me to display what the user types for their answer. Unfortunately, Java does not have this type of built in variable so I had to make my own, and it works well. I might run into some issues, but I will be able to implement it myself.

I was thinking going to use JavaFX for this project, but since I'm more skilled with AWT and Swing I will proceed.