✨ Added answer submission
Now you can submit your answer and if it is correct, then your total_marks will be added by the question_value. This will also increment your questions_solved, which is a variable used for progression.

Working on adding images so I can work on the upgrades system. Need to figure out how to do a scrolling area for all the upgrades. A new class will be required for the upgrades, since they are multiple instances of similar items with different properties.

I also commented more code and spaced it out for readability.