✨ Upgrade system progress
Implemented the upgrade creation system. So when the player reaches a certain amount of points, the corresponding upgrade will be displayed.

Created a new class called upgrade, which will be used to create the different types of upgrades.

Using the BufferedImage library to create the image of the button for the upgrade, this will be used for the upgrade icon as well.