✨ Upgrades fundamental functionality
Implemented functionality to click on the upgrade, added more attributes to the upgrade constructor like description, cost, and benefits. Icon images for the upgrades added to the project files, using ImageIO to display on the upgrade.

GameWindow now has a custom icon I made using Photoshop, instead of the default Java coffee icon.